Input requirement matrix

Returns input requirement matrix after passing an input-output matrix and a final demand vector as input.

input_requirement_matrix(X,d)

Arguments

X
input-output matrix

d
final demand vector

Value

input_requirement_matrix() returns a matrix, when the X matrix is not a square matrix or the d vector has a different number of elements than the number of rows in Xt it returns an error.

References

Suahasil Nazara & Dong Guo & Geoffrey J.D. Hewings & Chokri Dridi, 2004. "PyIO: Input-Output Analysis with Python," Urban/Regional 0409002, University of Illinois at Urbana-Champaign, USA.

Examples

  set.seed(200100)
  X <- matrix(rnorm(100), nrow = 10)
  d <- rnorm(10)
  input_requirement_matrix(X,d)

#>              [,1]       [,2]        [,3]       [,4]       [,5]        [,6]
#>  [1,]   9.5375187   4.596879  1.87979783 -0.1901464 -1.1402063  0.71885552
#>  [2,]  -8.1179649   5.648179 -2.20112083  1.5548465  0.8728831 -0.39416656
#>  [3,]  -6.7500630 -11.412299  0.75848651  0.6965183 -1.1721070  0.04963223
#>  [4,]  13.1738157  -4.881135 -2.24536599  2.5423619 -0.3966139  2.14909426
#>  [5,]   0.8465662   4.856727  0.09550012 -5.7980369  0.5113761  0.24975884
#>  [6,]  18.7758366   3.131802  1.64692938  0.9398365  0.5414045  0.82914379
#>  [7,] -25.7753279   2.400190 -1.77135771  3.8614778 -0.9632455 -1.21435042
#>  [8,]  -7.4562870  -2.687938  2.60594852  4.1574730  1.1474541  1.59813003
#>  [9,]  13.2548225   4.985968  1.03003782 -3.3012572  2.4439488 -1.42265527
#> [10,] -10.1381106   3.817424  2.98482296  0.7786192 -0.2581422 -1.93328597
#>             [,7]        [,8]       [,9]      [,10]
#>  [1,] -3.4838116 -0.67944855  0.5113307 -0.7628574
#>  [2,] -0.9360992 -0.32805097 -0.4932803  1.0935617
#>  [3,] -1.8589427  0.87297225 -0.1949332  1.2806116
#>  [4,]  0.9692820  1.51748699 -1.3267918 -2.4737694
#>  [5,]  1.9224045 -1.28913965  0.4061643  0.4169137
#>  [6,]  2.6215419  0.02239990 -2.4607614 -2.0180837
#>  [7,]  2.1935621  0.36130185  0.6407773  0.3118951
#>  [8,] -1.9175141 -1.27800392 -3.0426691 -1.0666533
#>  [9,]  1.3477232  0.00471119 -0.8704988  0.8714513
#> [10,]  2.2365196 -1.58967365  0.3798862 -0.9359380