Leontief inverse matrix

Returns leontief inverse matrix after passing an input requirement matrix as input.

leontief_inverse(A)

Arguments

A
input requirement matrix

Value

leontief_inverse() returns a matrix, when the A matrix is not square or the Leontief matrix can't be inverted it returns an error.

References

Suahasil Nazara & Dong Guo & Geoffrey J.D. Hewings & Chokri Dridi, 2004. "PyIO: Input-Output Analysis with Python," Urban/Regional 0409002, University of Illinois at Urbana-Champaign, USA.

Examples

  set.seed(200100)
  A <- matrix(rnorm(100), nrow = 10)
  leontief_inverse(A)

#>              [,1]        [,2]        [,3]        [,4]        [,5]        [,6]
#>  [1,]  0.16608344 -0.28552763 -0.11104490 -0.50488417 -0.08687074  0.11265454
#>  [2,]  0.36756775  0.06653501 -0.06070046  0.34561769  0.05996698 -0.28137088
#>  [3,] -0.67278421  0.45534960  0.01140609 -0.73657367  0.03216514  0.81261548
#>  [4,]  0.04008993  0.20036738  0.06704616  0.57037921 -0.08970220 -0.33174426
#>  [5,]  0.40439215  0.19967879  0.52353527  0.41793607  0.34923580  0.07674034
#>  [6,] -0.66825707  0.35163553  0.11844334 -1.45109077 -0.32173316  1.50304860
#>  [7,]  0.02068690 -0.19757168 -0.07903157  0.66213214  0.21077697 -0.54158586
#>  [8,] -0.67174598  0.39799978 -0.02850860 -0.47377786  0.12924743  0.55104416
#>  [9,]  0.21939357  0.06789227  0.33336164 -0.02410646  0.14913726  0.23542283
#> [10,]  0.48970149  0.14127608  0.39854292  1.61721106  0.29249859 -1.47522272
#>              [,7]        [,8]        [,9]       [,10]
#>  [1,]  0.53149615  0.21227308  0.02059626 -0.66533484
#>  [2,]  0.08942785 -0.08359409  0.36053774  0.21396691
#>  [3,] -0.15559372  0.04486814 -0.90431298 -0.62004311
#>  [4,] -0.20175890 -0.08302524  0.13543948  0.66402784
#>  [5,] -0.02253168 -0.17898180  0.04512023  0.44096318
#>  [6,] -0.07932259 -0.03029504 -1.19260699 -0.80329579
#>  [7,]  0.08243197 -0.05306084  0.47680473  0.53797869
#>  [8,] -0.40676662  0.13993726 -0.84047462 -0.03028693
#>  [9,]  0.22344549  0.14009254  0.24730262 -0.21395448
#> [10,] -0.29024271 -0.23800269  1.24765274  1.54390019