Output allocation matrix

Returns output allocation matrix after passing an input-output matrix and a final demand vector as input.

output_allocation_matrix(X,d)

Arguments

X
input-output matrix

d
final demand vector

Value

output_allocation_matrix() returns a matrix, when the X matrix is not a square matrix or the d vector has a different number of elements than the number of rows in X it returns an error.

References

Suahasil Nazara & Dong Guo & Geoffrey J.D. Hewings & Chokri Dridi, 2004. "PyIO: Input-Output Analysis with Python," Urban/Regional 0409002, University of Illinois at Urbana-Champaign, USA.

Examples

  set.seed(200100)
  X <- matrix(rnorm(100), nrow = 10)
  d <- rnorm(10)
  output_allocation_matrix(X,d)

#>             [,1]        [,2]        [,3]       [,4]       [,5]        [,6]
#>  [1,]  9.5375187 -11.9683156 11.08646581  0.8436806 -7.0802684  4.87195055
#>  [2,]  3.1180082   5.6481793  4.98604211  2.6497659 -2.0818643  1.02605614
#>  [3,] -1.1445265   5.0380340  0.75848651 -0.5240111 -1.2341034  0.05703522
#>  [4,] -2.9690785  -2.8641835  2.98455242  2.5423619  0.5550656 -3.28266691
#>  [5,]  0.1363310  -2.0363261  0.09070258  4.1429015  0.5113761  0.27259372
#>  [6,]  2.7703717  -1.2031036  1.43316335 -0.6152915  0.4960516  0.82914379
#>  [7,]  5.7419059   1.3920901  2.32723436  3.8167641  1.3324630  1.83339969
#>  [8,] -0.7508049   0.7046829  1.54757714 -1.8574787  0.7174739  1.09063107
#>  [9,]  1.8801338  -1.8413410  0.86168787  2.0777060  2.1526490 -1.36765252
#> [10,]  1.5211348   1.4912518 -2.64126043  0.5183531  0.2405116  1.96592990
#>             [,7]         [,8]       [,9]      [,10]
#>  [1,] 15.6387774 -6.747642737  3.6048489  5.0843181
#>  [2,] -1.6139872  1.251315734  1.3357002  2.7993855
#>  [3,]  1.4149209  1.469988589 -0.2330177 -1.4471874
#>  [4,]  0.9806372 -3.396491835  2.1081332 -3.7158540
#>  [5,] -1.3897178 -2.061717662  0.4611271 -0.4474755
#>  [6,] -1.7363756  0.032823156 -2.5597256  1.9845738
#>  [7,]  2.1935621 -0.799314258 -1.0063387  0.4630733
#>  [8,]  0.8667447 -1.278003915 -2.1599535  0.7158419
#>  [9,] -0.8581509  0.006636528 -0.8704988 -0.8238483
#> [10,]  1.5063695  2.368722308 -0.4018365 -0.9359380